feat!: simplify sdk usage by combining the views
Combined SmartSpectraButton and SmartSpectraResultView into a single SmartSpectraView
