+++ internal-demo/src/main/res/layout/activity_main.xml
@@ -17,14 +17,8 @@
             android:orientation="vertical"
             android:background="@android:color/white">
 
-            <com.presagetech.smartspectra.SmartSpectraButton
-                android:id="@+id/btn"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_margin="32dp" />
-
-            <com.presagetech.smartspectra.SmartSpectraResultView
-                android:id="@+id/result_view"
+            <com.presagetech.smartspectra.SmartSpectraView
+                android:id="@+id/smart_spectra_view"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"/>
