+++ sdk/src/main/java/com/presagetech/smartspectra/SmartSpectraButton.kt
@@ -4,7 +4,6 @@ import android.app.Dialog
 import android.content.Context
 import android.content.Intent
 import android.net.Uri
-import android.os.Build
 import android.util.AttributeSet
 import android.view.Gravity
 import android.view.LayoutInflater
@@ -16,7 +15,6 @@ import android.widget.Toast
 import androidx.appcompat.widget.AppCompatTextView
 import androidx.core.content.ContextCompat
 import com.google.android.material.bottomsheet.BottomSheetDialog
-import com.presage.physiology.proto.MetricsProto.MetricsBuffer
 import com.presagetech.smartspectra.ui.OnboardingTutorialActivity
 import com.presagetech.smartspectra.ui.SmartSpectraActivity
 import com.presagetech.smartspectra.ui.viewmodel.ScreeningViewModel
@@ -64,32 +62,6 @@ class SmartSpectraButton(context: Context, attrs: AttributeSet?) : LinearLayout(
 
         infoButton = findViewById(R.id.button_info)
         infoButton.setOnClickListener { infoBottomSheetDialog.show() }
-
-        // In case of unsupported devices
-        if (!isSupportedAbi()) {
-            isEnabled = false
-            checkupButton.isEnabled = false
-            Toast.makeText(context, "Unsupported device (ABI)", Toast.LENGTH_LONG).show()
-            Timber.d("Unsupported device (ABI)")
-            Timber.d("This device ABIs: ${Build.SUPPORTED_ABIS.contentToString()}")
-        } else {
-            // Load necessary libraries
-            System.loadLibrary("mediapipe_jni")
-            System.loadLibrary("opencv_java3")
-        }
-
-    }
-
-    fun setApiKey(apiKey: String) {
-        screeningViewModel.setApiKey(apiKey)
-    }
-
-    fun setMeshPointsObserver(observer: (List<Pair<Int, Int>>) -> Unit) {
-        screeningViewModel.observeDenseMeshPoints(observer)
-    }
-
-    fun setMetricsBufferObserver(observer: (MetricsBuffer) -> Unit) {
-        screeningViewModel.observeMetricsBuffer(observer)
     }
 
     private val infoBottomSheetDialog: BottomSheetDialog by lazy {
@@ -245,22 +217,5 @@ class SmartSpectraButton(context: Context, attrs: AttributeSet?) : LinearLayout(
     private fun dpToPx(dp: Int): Int {
         val density = context.resources.displayMetrics.density
         return (dp * density).roundToInt()
-    }
-
-    fun setSpotTime(spotDuration: Double) {
-        SmartSpectraSDKConfig.spotDuration = spotDuration
-    }
-
-    fun setShowFps(showFps: Boolean) {
-        SmartSpectraSDKConfig.SHOW_FPS = showFps
-    }
-
-    private fun isSupportedAbi(): Boolean {
-        Build.SUPPORTED_ABIS.forEach {
-            if (it == "arm64-v8a" || it == "armeabi-v7a") {
-                return true
-            }
-        }
-        return false
     }
 }
